+++ src/main/java/com/grilledsausage/molva/api/dto/user/OAuthToken.java
@@ -1,4 +1,4 @@
-package com.grilledsausage.molva.api.dto;
+package com.grilledsausage.molva.api.dto.user;
 
 import lombok.Data;
